# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: 企业内部练习与考试系统 API
  version: 1.0.0
  description: |
    MVP 完整接口规格。全局约定见 [00-API设计说明.md](./00-API设计说明.md)。
    分域行为见 01~07 Markdown 文档。
  contact:
    name: Exam System Team
servers:
  - url: /api/v1
    description: 单租户企业内网 API
...
security:
  - sessionAuth: []
paths:
...
  /departments:
    get:
      tags: [Organization]
      operationId: listDepartments
      summary: 部门列表/树
      x-fr-ids: [FR-AUTH-04]
      x-page-ids: [AD-02]
      parameters:
        - name: format
          in: query
          schema:
            type: string
            enum: [tree, flat]
            default: tree
      responses:
        '200':
          description: 部门数据
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Department'
    post:
      tags: [Organization]
      operationId: createDepartment
      summary: 创建部门
      x-fr-ids: [FR-AUTH-04]
      x-page-ids: [AD-02]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, parentId]
              properties:
                name:
                  type: string
                parentId:
                  type: string
      responses:
        '201':
          description: 已创建
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Department'
  /departments/{id}:
    patch:
      tags: [Organization]
      operationId: updateDepartment
      summary: 更新部门（改名/移动/停用）
      x-fr-ids: [FR-AUTH-04]
      x-page-ids: [AD-02]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                parentId:
                  type: string
                status:
                  type: string
                  enum: [active, disabled]
      responses:
        '200':
          $ref: '#/components/responses/SuccessEmpty'
  /employees:
    get:
      tags: [Organization]
      operationId: listEmployees
      summary: 员工分页列表
      x-fr-ids: [FR-AUTH-04]
      x-page-ids: [AD-03]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: departmentId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [active, disabled]
        - name: keyword
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 分页员工
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PagedEmployees'
    post:
      tags: [Organization]
      operationId: createEmployee
      summary: 手工建档
      x-fr-ids: [FR-AUTH-04]
      x-page-ids: [AD-03]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEmployeeRequest'
      responses:
        '201':
          description: 含一次性临时密码
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CreateEmployeeResponse'
...
  /employees/{id}:
    get:
      tags: [Organization]
      operationId: getEmployee
      summary: 员工详情
      x-fr-ids: [FR-AUTH-04]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          $ref: '#/components/responses/SuccessEmpty'
    patch:
      tags: [Organization]
      operationId: updateEmployee
      summary: 更新员工
      x-fr-ids: [FR-AUTH-04]
      x-page-ids: [AD-03, AD-04]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                displayName:
                  type: string
                departmentPath:
                  type: string
                phone:
                  type: string
                status:
                  type: string
                  enum: [active, disabled]
      responses:
        '200':
          $ref: '#/components/responses/SuccessEmpty'
  /employees/{id}/reset-password:
    post:
      tags: [Organization]
      operationId: resetEmployeePassword
      summary: 重置员工密码
      x-fr-ids: [FR-AUTH-02]
      x-page-ids: [AD-04]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: 含一次性临时密码
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          temporaryPassword:
                            type: string
...
  /employees/{id}/admin-grants:
    patch:
      tags: [Organization]
      operationId: updateAdminGrants
      summary: 管理员与异常处置授权
      x-fr-ids: [FR-AUTH-05]
      x-page-ids: [AD-04]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                isAdmin:
                  type: boolean
                hasOutageDisposition:
                  type: boolean
                reason:
                  type: string
      responses:
        '200':
          $ref: '#/components/responses/SuccessEmpty'
...
  /practice/sessions:
    post:
      tags: [Practice]
      operationId: createPracticeSession
      summary: 开始练习
      x-fr-ids: [FR-PRA-02, FR-PRA-03]
      x-page-ids: [MP-03, MP-04]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePracticeSessionRequest'
      responses:
        '201':
          $ref: '#/components/responses/SuccessEmpty'
        '409':
          $ref: '#/components/responses/Conflict'
  /practice/sessions/{id}:
    get:
      tags: [Practice]
      operationId: getPracticeSession
      summary: 练习会话详情
      x-fr-ids: [FR-PRA-02]
      x-page-ids: [MP-04]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          $ref: '#/components/responses/SuccessEmpty'
  /practice/sessions/{id}/answers:
    post:
      tags: [Practice]
      operationId: submitPracticeAnswer
      summary: 逐题提交（即时反馈）
      x-fr-ids: [FR-PRA-04]
      x-page-ids: [MP-04]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [questionVersionId, answer]
              properties:
                questionVersionId:
                  type: string
                answer:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          $ref: '#/components/responses/SuccessEmpty'
  /practice/sessions/{id}/finish:
    post:
      tags: [Practice]
      operationId: finishPracticeSession
      summary: 结束练习
      x-fr-ids: [FR-PRA-02]
      x-page-ids: [MP-04]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          $ref: '#/components/responses/SuccessEmpty'
  /practice/wrong-book:
    get:
      tags: [Practice]
      operationId: listWrongBook
      summary: 错题本
      x-fr-ids: [FR-PRA-04, FR-PRA-05]
      x-page-ids: [MP-05]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          $ref: '#/components/responses/SuccessEmpty'
  /practice/records:
    get:
      tags: [Practice]
      operationId: listPracticeRecords
      summary: 练习记录
      x-fr-ids: [FR-PRA-05]
      x-page-ids: [MP-11]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          $ref: '#/components/responses/SuccessEmpty'
...
  /admin/outage-events:
    get:
      tags: [Outage]
      operationId: listOutageEvents
      summary: 故障事件列表
      x-fr-ids: [FR-OPS-01]
      x-page-ids: [AD-13]
      x-sys-ids: [SYS-05]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          $ref: '#/components/responses/SuccessEmpty'
  /admin/outage-events/{id}:
    get:
      tags: [Outage]
      operationId: getOutageEvent
      summary: 故障事件详情与提案
      x-fr-ids: [FR-OPS-01, FR-OPS-02]
      x-page-ids: [AD-13]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          $ref: '#/components/responses/SuccessEmpty'
  /admin/outage-events/{id}/proposals/{version}/confirm:
    post:
      tags: [Outage]
      operationId: confirmOutageProposal
      summary: 确认故障提案并补时
      x-fr-ids: [FR-OPS-02]
      x-page-ids: [AD-13]
      x-sys-ids: [SYS-05]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - name: version
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                confirmationNote:
                  type: string
      responses:
        '200':
          $ref: '#/components/responses/SuccessEmpty'
        '409':
          $ref: '#/components/responses/Conflict'
  /admin/outage-events/{id}/proposals/{version}/reject:
    post:
      tags: [Outage]
      operationId: rejectOutageProposal
      summary: 驳回提案进入复核
      x-fr-ids: [FR-OPS-02]
      x-page-ids: [AD-13]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - name: version
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rejectReason]
              properties:
                rejectReason:
                  type: string
      responses:
        '200':
          $ref: '#/components/responses/SuccessEmpty'
...
components:
  securitySchemes:
    sessionAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 亦可用 HttpOnly Session Cookie，二者等价
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  responses:
    SuccessEmpty:
      description: 成功
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiEnvelope'
...
    Conflict:
      description: 版本或状态冲突
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
    ApiEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: {}
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    ErrorEnvelope:
      type: object
      required: [error, meta]
      properties:
        error:
          $ref: '#/components/schemas/ApiError'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    ResponseMeta:
      type: object
      required: [serverNow, requestId]
      properties:
        serverNow:
          type: string
          format: date-time
        requestId:
          type: string
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    ErrorCode:
      type: string
      enum:
        - AUTH_INVALID_CREDENTIALS
        - AUTH_ACCOUNT_LOCKED
        - AUTH_ACCOUNT_DISABLED
        - AUTH_SESSION_EXPIRED
        - AUTH_PASSWORD_CHANGE_REQUIRED
        - AUTH_SMS_RATE_LIMITED
        - ORG_DEPARTMENT_NAME_DUPLICATE
        - ORG_LAST_OUTAGE_ADMIN
        - IMP_PREVIEW_STALE
        - PRA_SESSION_ALREADY_ACTIVE
        - SIM_ATTEMPT_ALREADY_ACTIVE
        - ATT_NO_REMAINING_OPPORTUNITY
        - ATT_EXAM_PAUSED
        - ANS_VERSION_CONFLICT
        - ANS_UNCONFIRMED_ANSWERS
        - ANS_ATTEMPT_TERMINATED
        - SCR_SUBMIT_IN_PROGRESS
        - OPS_PROPOSAL_STALE
        - OPS_NOT_AUTHORIZED
        - SEC_FORBIDDEN
...
    Department:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        parentId:
          type: [string, 'null']
        path:
          type: string
        status:
          type: string
          enum: [active, disabled]
        employeeCount:
          type: integer
        children:
          type: array
          items:
            $ref: '#/components/schemas/Department'
    EmployeeSummary:
      type: object
      properties:
        id:
          type: string
        employeeNo:
          type: string
        displayName:
          type: string
        departmentPath:
          type: string
        phoneMasked:
          type: string
        status:
          type: string
          enum: [active, disabled]
        isAdmin:
          type: boolean
        hasOutageDisposition:
          type: boolean
    PagedEmployees:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/EmployeeSummary'
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
    CreateEmployeeRequest:
      type: object
      required: [employeeNo, displayName, departmentPath]
      properties:
        employeeNo:
          type: string
        displayName:
          type: string
        departmentPath:
          type: string
        phone:
          type: string
    CreateEmployeeResponse:
      type: object
      properties:
        employee:
          $ref: '#/components/schemas/EmployeeSummary'
        temporaryPassword:
          type: string
          description: 仅本次响应返回一次
...
    CreatePracticeSessionRequest:
      type: object
      required: [questionBankId, mode]
      properties:
        questionBankId:
          type: string
        mode:
          type: string
          enum: [sequential, random, targeted, wrongBook]
        scope:
          type: object
          properties:
            categoryId:
              type: [string, 'null']
            knowledgePointId:
              type: [string, 'null']
        questionCount:
          type: integer
          enum: [10, 20, 50]
...